refactor: extract bookrepository to adapters file
